Try to avoid connecting to DB if password missing
For some installs of RT, too many failed database connections will lock
account access. Regardless of whether this is a permission error,
configuration error, etc., try to mitigate this by not attempting to
connect if the database password is blank or missing.

Add a configuration option, AllowEmptyDatabasePassword, to skip the
requirement of having a non-empty database password to start RT.